refactor: soft delete with used_at, all price is int
- use promotion results in soft delete (used_at) in DB.
- BigDecimal calculation reserved to only important discount calculation. All price representation is int now

+++ src/main/java/antigravity/model/response/ProductAmountResponse.java
@@ -3,13 +3,11 @@ package antigravity.model.response;
 import lombok.Builder;
 import lombok.Data;
 
-import java.math.BigDecimal;
-
 @Data
 @Builder
 public class ProductAmountResponse {
     private String name; //상품명
     private int originPrice; //상품 기존 가격
-    private BigDecimal discountPrice; //총 할인 금액
+    private int discountPrice; //총 할인 금액
     private int finalPrice; //확정 상품 가격
 }
